(BSR)[API] feat: add time measure on each industrial sandbox step
