+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_closed_offerers.py
@@ -17,11 +17,13 @@ from pcapi.core.offers import factories as offers_factories
 from pcapi.core.users import factories as users_factories
 from pcapi.core.users import models as users_models
 from pcapi.models import db
+from pcapi.sandboxes.scripts.utils.helpers import log_func_duration
 
 
 logger = logging.getLogger(__name__)
 
 
+@log_func_duration
 def create_closed_offerers() -> None:
     logger.info("create_closed_offerers")
