+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_industrial_bank_accounts.py
@@ -4,11 +4,13 @@ from pcapi.core.finance import factories as finance_factories
 from pcapi.core.finance import models as finance_models
 from pcapi.core.offerers import factories as offerers_factories
 from pcapi.core.offers import factories as offers_factories
+from pcapi.sandboxes.scripts.utils.helpers import log_func_duration
 
 
 logger = logging.getLogger(__name__)
 
 
+@log_func_duration
 def create_industrial_bank_accounts() -> None:
     logger.info("create_industrial_bank_accounts")
     create_offerer_without_bank_accounts()
