+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_industrial_draft_offers.py
@@ -4,11 +4,13 @@ import pcapi.core.offerers.models as offerers_models
 import pcapi.core.offers.factories as offers_factories
 from pcapi.models import db
 from pcapi.models.offer_mixin import OfferValidationStatus
+from pcapi.sandboxes.scripts.utils.helpers import log_func_duration
 
 
 logger = logging.getLogger(__name__)
 
 
+@log_func_duration
 def create_industrial_draft_offers(offerers_by_name: dict[str, offerers_models.Offerer]) -> None:
     logger.info("create_industrial_individual_draft_offers")
     index = 0
