+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_industrial_eac_data/create_users_and_offerers.py
@@ -1,7 +1,9 @@
 from pcapi.core.offerers import factories as offerers_factories
 from pcapi.core.offerers import models as offerers_models
+from pcapi.sandboxes.scripts.utils.helpers import log_func_duration
 
 
+@log_func_duration
 def create_users_offerers() -> list[offerers_models.Offerer]:
     # WARNING: please add offerers only at the end, because the order in the returned list is important to keep
     # consistency in create_industrial_eac_data.create_venues.create_venues
