+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_industrial_individual_offerers.py
@@ -10,11 +10,13 @@ from pcapi.core.offerers import factories as offerers_factories
 from pcapi.core.offerers import models as offerers_models
 from pcapi.core.users import factories as users_factories
 from pcapi.models import db
+from pcapi.sandboxes.scripts.utils.helpers import log_func_duration
 
 
 logger = logging.getLogger(__name__)
 
 
+@log_func_duration
 def create_industrial_individual_offerers() -> None:
     logger.info("create_industrial_individual_offerers")
