+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_industrial_search_objects.py
@@ -6,11 +6,13 @@ from pcapi.core.educational import models as educational_models
 from pcapi.core.offerers.models import Venue
 from pcapi.core.offers.models import Offer
 from pcapi.models import db
+from pcapi.sandboxes.scripts.utils.helpers import log_func_duration
 
 
 logger = logging.getLogger(__name__)
 
 
+@log_func_duration
 def create_industrial_search_indexed_objects() -> None:
     if settings.ALGOLIA_TRIGGER_INDEXATION:
         logger.info("create_industrial_algolia_objects")
