+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_industrial_user_tags.py
@@ -1,11 +1,13 @@
 import logging
 
 from pcapi.core.users import factories as users_factories
+from pcapi.sandboxes.scripts.utils.helpers import log_func_duration
 
 
 logger = logging.getLogger(__name__)
 
 
+@log_func_duration
 def create_industrial_user_tags() -> None:
     logger.info("create_industrial_user_tags")
