+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_industrial_venues_with_timezone.py
@@ -1,11 +1,13 @@
 import logging
 
 from pcapi.core.offerers import factories as offerers_factories
+from pcapi.sandboxes.scripts.utils.helpers import log_func_duration
 
 
 logger = logging.getLogger(__name__)
 
 
+@log_func_duration
 def create_industrial_venues_with_timezone() -> None:
     logger.info("create_industrial_venues_with_timezone")
     offerer = offerers_factories.OffererFactory.create(name="Structure avec différentes timezones")
