+++ api/src/pcapi/sandboxes/scripts/creators/industrial/create_user_offerers.py
@@ -4,11 +4,13 @@ from pcapi.core.offerers import factories as offerers_factories
 from pcapi.core.offerers import models as offerers_models
 from pcapi.core.users import factories as users_factories
 from pcapi.models import db
+from pcapi.sandboxes.scripts.utils.helpers import log_func_duration
 
 
 logger = logging.getLogger(__name__)
 
 
+@log_func_duration
 def create_user_offerers() -> None:
     # Add a user and user_offerer for each offerer without one, except if the offerer has a provider
     logger.info("create_user_offerers")
